Less Work done for the same Effect
Previously, the immediate speed was computed regardless of
wheteher it was going to be displayed or not. After this
change, it is not going to be computed unless it is going to
be displayed. Also, it won't be computed either if the re-
gistered travel distance is too small, it will be just set
to zero.