Optimized Tracking of the Total Trailing Speed Length an Time
This change removes the necessity to compute the sum
of all stored data points on position and time by iterating
through all of them. Instead, it is keeping the suum
updated by always adding the value that has come in, and
removing the other value that has left the data set.

Therefore most of the totalTime and totalLength tracking
has been moved to the data managing section.

Also, the number of stored points has been increased to 25.