The immediate speed is gone now
Following the previous changed, the speed computed based
on the locally stored data set has become so sufficient
for the purposes of general use that this speed computed
between the two latest data points has become unnecessary
to compute and present to users.

Therefore, saying goodbye at this point. It can be easily
re-added if necessary.