Added Scaling based on Display Resolution
So far, TCDM has only been rendering well on 1080p. It turns out,
any other resolution would prove it to be too big or too small.

This change works toward taking user's display resolution into
account when rendering the TCDM window. To avoid issues when the
user is trying out things, the resolution is checked every 79
frames or so, and is also recomputed on resetting TCDM.