Display Speed Periodically
So far, the raw speed reading was making it into the speedometer,
which sometimes resulted in quite jumpy results that may be im-
possible to read. This change introduces the concept of displaying
the output speed at regular time intervals, which are programmable.

Presently, things are set to display the speed reading every 0.15
seconds.