TCDM: Quicker Update Rate for the Car's Speedometer
This change lets the game decide when to display the new speed reading.

So far, TCDM has been having a new speed computation on every 2nd frame,
but it has been controlling the update of the value in the UI, with an
assumed benefit for the stability of the displayed number. As much as
that may be true, it is not what the drivers want. They appreciate get-
ting speed updates as they happen. Hence, the change.